Send heartbeat to the master
